Remove client from sawtooth.sdk.client.signing package
The package now is sawtooth.sdk.signing.
