+++ sawtooth-sdk-signing/src/main/java/sawtooth/sdk/signing/Secp256k1PrivateKey.java
@@ -12,7 +12,7 @@
  limitations under the License.
 ------------------------------------------------------------------------------*/
 
-package sawtooth.sdk.client.signing;
+package sawtooth.sdk.signing;
 
 
 import org.bitcoinj.core.Utils;
